requirements.txt
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
coloredlogs==15.0.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
ftfy==6.3.1
huggingface-hub==0.32.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
open_clip_torch==2.32.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pathlib==1.0.1
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.0
PyMatting==1.1.14
pyreadline3==3.5.4
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.32.3
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.8.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.24
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
